requirements.txt
mysql-connector-python
numpy
pandas
matplotlib
seaborn
plotly
streamlit
streamlit>=1.32.0
altair<5
